refactor(discriminator): move DiscriminatorTypeComposer to own file
