+++ src/discriminators/index.js
@@ -1,3 +1,4 @@
 /* @flow */
 
-export { composeChildTC } from './composeChildTC';
+export { DiscriminatorTypeComposer, Options } from './DiscriminatorTypeComposer';
+export { mergeCustomizationOptions } from './merge-customization-options';
