+++ src/discriminators/prepare-resolvers/prepareChildResolvers.js
@@ -2,8 +2,7 @@
 
 import type { ResolveParams } from 'graphql-compose';
 import { Resolver, TypeComposer } from 'graphql-compose';
-import type { Options } from '../../composeWithMongooseDiscriminators';
-import { DiscriminatorTypeComposer } from '../../composeWithMongooseDiscriminators';
+import { type Options, DiscriminatorTypeComposer } from '../DiscriminatorTypeComposer';
 import { EMCResolvers } from '../../resolvers';
 
 // set the DKey as a query on filter, also project it
